Moved routes into server root folder. cleaner folder strucutre
