+++ server/routes/publicRoute.js
@@ -1,7 +1,6 @@
 'use strict';
-var publicController = require('../../controller/publicController')
+var publicController = require('../controller/publicController')
 
 module.exports = function (app) {
-  // app.get('/', publicController.rootPath);
   app.get('/download/:snippetId', publicController.downloadSnippet);
 }
