+++ server/routes/userRoute.js
@@ -1,6 +1,6 @@
 'use strict';
 // /api/user route
-var userController = require('../../controller/userController');
+var userController = require('../controller/userController');
 
 module.exports = function (app) {
   app.get('/:username', userController.userPage);
